// ClientApp/src/app/models/docente.ts
export class Docente {
    id: number;
    identificacion: string;
    primerNombre: string;
    segundoNombre: string;
    primerApellido: string;
    segundoApellido: string;
    correo: string;
    //fechaNacimiento: string;
    genero: string;
    //telefono: number;
    estado: string;
    cargo: string;
    programa: string;
    facultad: string;
}